+++ VibeMeter/App/VibeMeterApp.swift
@@ -2,29 +2,6 @@ import AppKit
 import os.log
 import SwiftUI
 
-// MARK: - Environment Detection Helper
-
-extension ProcessInfo {
-    /// Detects if the app is running in SwiftUI previews
-    var isRunningInPreview: Bool {
-        environment["XCODE_RUNNING_FOR_PREVIEWS"] == "1"
-    }
-    
-    /// Detects if the app is running in test environment
-    var isRunningInTests: Bool {
-        environment["XCTestConfigurationFilePath"] != nil || NSClassFromString("XCTestCase") != nil
-    }
-    
-    /// Detects if running in debug mode (using DEBUG compilation condition)
-    var isRunningInDebug: Bool {
-        #if DEBUG
-        return true
-        #else
-        return false
-        #endif
-    }
-}
-
 // MARK: - App Entry Point
 
 /// The main entry point for the VibeMeter application.
